requirements.txt
fastapi
uvicorn
pydantic
pyyaml
requests
pymongo